+++ nossl/random.cpp
@@ -55,7 +55,7 @@ bool Random::seed(const unsigned char *buf, size_t size)
 size_t Random::key(unsigned char *buf, size_t size)
 {
 #ifdef  _MSWINDOWS_
-    if((__handle != (HCRYPTPROV)NULL) && CryptGenRandom(__handle, size, buf))
+    if((__handle != (HCRYPTPROV)NULL) && CryptGenRandom(__handle, (DWORD)size, buf))
         return size;
     return 0;
 #else
